Fix issue that prevented keyboard layout from applying during Calamares installation
The calamares-nixos package has a fix for setting the keyboard layout through gsettings but this patch is only applied in the code path used when Calamares is not relying on systemd's localed to set the keymap. Configuring Calamares to not use locale1 forces it to go through the code path that uses xkbmap and that has the nixos specific patch.
This can probably be done more cleanly upstream but we're not sure we'll keep using the Calamares installer for deskOS anyway.